Fix bugs and record meta-info

+++ main.py
@@ -14,6 +14,8 @@ import subprocess
 import traceback
 from typing import Iterator, List, NamedTuple, Optional
 
+from mypy_extensions import TypedDict
+
 import ghcc
 from ghcc import CloneErrorType, RepoMakefileEntry
 
@@ -41,6 +43,8 @@ def parse_args():
     parser.add_argument("--max-repos", type=int,
                         default=-1)  # maximum number of repositories to process (ignoring non-existent)
     parser.add_toggle_argument("--recursive-clone", default=True)  # if True, use `--recursive` when `git clone`
+    parser.add_toggle_argument("--write-db", default=True)  # only modify the DB when True
+    parser.add_toggle_argument("--record-metainfo", default=False)  # if True, record a bunch of other stuff
     # parser.add_argument("--automake", action="store_true", default=False)  # if True, support automake
     args = parser.parse_args()
     return args
@@ -53,13 +57,19 @@ class RepoInfo(NamedTuple):
     db_result: Optional[ghcc.RepoEntry]
 
 
+class PipelineMetaInfo(TypedDict):
+    r"""Meta-info that might be required for experimentations."""
+    num_makefiles: int  # total number of Makefiles
+    has_gitmodules: bool  # whether repo contains a .gitmodules file
+
+
 class PipelineResult(NamedTuple):
-    repo_owner: str
-    repo_name: str
+    repo_info: RepoInfo
     clone_success: Optional[bool] = None
     repo_size: Optional[int] = None
     makefiles: Optional[List[RepoMakefileEntry]] = None
     libraries: Optional[List[str]] = None
+    meta_info: Optional[PipelineMetaInfo] = None
 
 
 def contains_in_file(file_path: str, text: str) -> bool:
@@ -99,11 +109,15 @@ def _docker_batch_compile(index: int, repo_binary_dir: str, repo_path: str,
 def exception_handler(e, *args, **kwargs):
     repo_info: RepoInfo = args[0] if len(args) > 0 else kwargs["repo_info"]
     exc_msg = f"<{e.__class__.__qualname__}> {e}"
-    if not (isinstance(e, subprocess.CalledProcessError) and e.output is not None):
-        ghcc.log(traceback.format_exc(), "error")
+    try:
+        if not (isinstance(e, subprocess.CalledProcessError) and e.output is not None):
+            ghcc.log(traceback.format_exc(), "error")
 
-    ghcc.log(f"Exception occurred when processing {repo_info.repo_owner}/{repo_info.repo_name}: {exc_msg}", "error")
-    # raise e
+        ghcc.log(f"Exception occurred when processing {repo_info.repo_owner}/{repo_info.repo_name}: {exc_msg}", "error")
+    except Exception as log_e:
+        print(f"Exception occurred when processing {repo_info.repo_owner}/{repo_info.repo_name}: {exc_msg}")
+        print(f"Another exception occurred while logging: <{log_e.__class__.__qualname__}> {log_e}")
+        raise log_e
 
 
 @ghcc.utils.exception_wrapper(exception_handler)
@@ -112,7 +126,7 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
                       clone_timeout: Optional[float] = None, compile_timeout: Optional[float] = None,
                       force_reclone: bool = False, force_recompile: bool = False, docker_batch_compile: bool = True,
                       max_archive_size: Optional[int] = None, compression_type: str = "gzip",
-                      record_libraries: bool = False) -> PipelineResult:
+                      record_libraries: bool = False, record_metainfo: bool = False) -> PipelineResult:
     r"""Perform the entire pipeline.
 
     :param repo_info: Information about the repository.
@@ -137,6 +151,7 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
     :param compression_type: The file type of the archive to produce. Valid values are ``"gzip"`` (faster) and
         ``"xz"`` (smaller).
     :param record_libraries: If ``True``, record the libraries used in compilation.
+    :param record_metainfo: If ``True``, record meta-info values.
 
     :return: An entry to insert into the DB, or `None` if no operations are required.
     """
@@ -144,28 +159,34 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
     repo_folder_name = f"{repo_info.repo_owner}_____{repo_info.repo_name}"
     repo_path = os.path.join(clone_folder, repo_folder_name)
     if compression_type == "xz":
-        archive_path = os.path.join(archive_folder, f"{repo_full_name}.tar.xz")
+        archive_extension = ".tar.xz"
         tar_type_flag = "J"
     elif compression_type == "gzip":
-        archive_path = os.path.join(archive_folder, f"{repo_full_name}.tar.gz")
+        archive_extension = ".tar.gz"
         tar_type_flag = "z"
     else:
         raise ValueError(f"Invalid compression type '{compression_type}'")
+    archive_path = os.path.abspath(os.path.join(archive_folder, f"{repo_full_name}{archive_extension}"))
 
     repo_entry = repo_info.db_result
-    repo_size = None
     clone_success = None
+
+    # Skip repos that are fully processed
+    if (repo_entry is not None and
+            (repo_entry["clone_successful"] and not force_reclone) and
+            (repo_entry["compiled"] and not force_recompile)):
+        return PipelineResult(repo_info)
 
     # Stage 1: Cloning from GitHub.
     if not force_reclone and os.path.exists(archive_path):
         # Extract the archive instead of cloning.
         try:
-            ghcc.utils.run_command(["tar", f"x{tar_type_flag}f", archive_path, repo_path])
+            ghcc.utils.run_command(["tar", f"x{tar_type_flag}f", archive_path], timeout=clone_timeout, cwd=clone_folder)
             ghcc.log(f"{repo_full_name} extracted from archive", "success")
         except (subprocess.TimeoutExpired, subprocess.CalledProcessError) as e:
             ghcc.log(f"Unknown error when extracting {repo_full_name}. Captured output: '{e.output}'", "error")
             shutil.rmtree(repo_path)
-            return PipelineResult(repo_info.repo_owner, repo_info.repo_name)
+            return PipelineResult(repo_info)  # return dummy info
         repo_size = ghcc.utils.get_folder_size(repo_path)
     elif (repo_entry is None or  # not processed
           (repo_entry["clone_successful"] and  # not compiled
@@ -189,20 +210,21 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
                 ghcc.log(msg, "error")
 
                 if clone_result.error_type is CloneErrorType.Unknown:
-                    return PipelineResult(repo_info.repo_owner, repo_info.repo_name)  # return dummy info
+                    return PipelineResult(repo_info)  # return dummy info
 
-            return PipelineResult(repo_info.repo_owner, repo_info.repo_name, clone_success=clone_success)
+            return PipelineResult(repo_info, clone_success=clone_success)
 
         repo_size = ghcc.utils.get_folder_size(repo_path)
         ghcc.log(f"{repo_full_name} successfully cloned ({clone_result.time:.2f}s, "
                  f"{ghcc.utils.readable_size(repo_size)})", "success")
     else:
         if not repo_entry["clone_successful"]:
-            return PipelineResult(repo_info.repo_owner, repo_info.repo_name)  # return dummy info
+            return PipelineResult(repo_info)  # return dummy info
         repo_size = ghcc.utils.get_folder_size(repo_path)
 
     makefiles = None
     libraries = None
+    meta_info = None
     if not repo_entry or not repo_entry["compiled"] or force_recompile:
         # # SPECIAL CHECK: Do not attempt to compile OS kernels!
         # kernel_name = None
@@ -214,8 +236,7 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
         #     shutil.rmtree(repo_path)
         #     ghcc.log(f"Found {kernel_name} kernel in {repo_full_name}, will not attempt to compile. "
         #              f"Repository deleted", "warning")
-        #     return PipelineResult(repo_info.repo_owner, repo_info.repo_name,
-        #                           clone_success=clone_success, makefiles=[])
+        #     return PipelineResult(repo_info, clone_success=clone_success, makefiles=[])
 
         # Stage 2: Finding Makefiles.
         makefile_dirs = ghcc.find_makefiles(repo_path)
@@ -223,7 +244,7 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
             # Repo has no Makefiles, delete.
             shutil.rmtree(repo_path)
             ghcc.log(f"No Makefiles found in {repo_full_name}, repository deleted", "warning")
-            return PipelineResult(repo_info.repo_owner, repo_info.repo_name, clone_success=clone_success, makefiles=[])
+            return PipelineResult(repo_info, clone_success=clone_success, makefiles=[])
         else:
             pass
 
@@ -253,6 +274,12 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
               f"in {repo_full_name} compiled (partially), yielding {num_binaries} binaries"
         ghcc.log(msg, "success" if num_succeeded == len(makefile_dirs) else "warning")
 
+        if record_metainfo:
+            meta_info: PipelineMetaInfo = {
+                "num_makefiles": len(makefile_dirs),
+                "has_gitmodules": os.path.exists(os.path.join(repo_path, ".gitmodules")),
+            }
+
         # Stage 4: Clean and zip repo.
         if max_archive_size is not None and repo_size > max_archive_size:
             shutil.rmtree(repo_path)
@@ -263,7 +290,8 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
             os.makedirs(os.path.split(archive_path)[0], exist_ok=True)
             compress_success = False
             try:
-                ghcc.utils.run_command(["tar", f"c{tar_type_flag}f", archive_path, repo_path], timeout=clone_timeout)
+                ghcc.utils.run_command(["tar", f"c{tar_type_flag}f", archive_path, repo_folder_name],
+                                       timeout=clone_timeout, cwd=clone_folder)
                 compress_success = True
             except subprocess.TimeoutExpired:
                 ghcc.log(f"Compression timeout for {repo_full_name}, giving up", "error")
@@ -275,11 +303,11 @@ def clone_and_compile(repo_info: RepoInfo, clone_folder: str, binary_folder: str
             elif os.path.exists(archive_path):
                 os.remove(archive_path)
 
-    return PipelineResult(repo_info.repo_owner, repo_info.repo_name, clone_success=clone_success,
-                          repo_size=repo_size, makefiles=makefiles, libraries=libraries)
+    return PipelineResult(repo_info, clone_success=clone_success, repo_size=repo_size,
+                          makefiles=makefiles, libraries=libraries, meta_info=meta_info)
 
 
-def iter_repos(db: ghcc.Database, repo_list_path: str) -> Iterator[RepoInfo]:
+def iter_repos(db: ghcc.Database, repo_list_path: str, max_count: Optional[int] = None) -> Iterator[RepoInfo]:
     db_entries = {
         (entry["repo_owner"], entry["repo_name"]): entry
         for entry in db.collection.find()
@@ -296,6 +324,55 @@ def iter_repos(db: ghcc.Database, repo_list_path: str) -> Iterator[RepoInfo]:
             db_result = db_entries.get((repo_owner, repo_name), None)
             yield RepoInfo(index, repo_owner, repo_name, db_result)
             index += 1
+            if max_count is not None and index >= max_count:
+                break
+
+
+class MetaInfo:
+    def __init__(self):
+        self.num_repos = 0
+        self.num_makefiles = 0
+        self.num_binaries = 0
+        self.num_gitmodules = 0
+        self.fail_to_success = 0
+        self.success_to_fail = 0
+        self.success_makefiles = 0
+        self.added_makefiles = 0
+        self.missing_makefiles = 0
+
+    def add_repo(self, result: PipelineResult) -> None:
+        assert result.meta_info is not None
+        self.num_repos += 1
+        self.num_gitmodules += result.meta_info["has_gitmodules"]
+        self.num_makefiles += result.meta_info["num_makefiles"]
+
+        new_makefiles = {}
+        old_makefiles = {}
+        if result.makefiles is not None:
+            new_makefiles = {makefile["directory"]: makefile["success"]
+                             for makefile in result.makefiles}
+            self.success_makefiles += sum(new_makefiles.values())
+        if result.repo_info.db_result is not None:
+            old_makefiles = {makefile["directory"]: makefile["success"]
+                             for makefile in result.repo_info.db_result["makefiles"]}
+        new_makefile_set = set(new_makefiles.keys())
+        old_makefile_set = set(old_makefiles.keys())
+        self.added_makefiles += len(new_makefile_set.difference(old_makefile_set))
+        self.missing_makefiles += len(old_makefile_set.difference(new_makefile_set))
+        for name in new_makefile_set.intersection(old_makefile_set):
+            new_status = new_makefiles[name]
+            old_status = old_makefiles[name]
+            if new_status and not old_status:
+                self.fail_to_success += 1
+            elif old_status and not new_status:
+                self.success_to_fail += 1
+
+    def __repr__(self) -> str:
+        msg = f"#Repos: {self.num_repos} ({self.num_gitmodules} with .gitmodules), #Binaries: {self.num_binaries}" \
+              f"#Makefiles: {self.num_makefiles} ({self.success_makefiles} succeeded)\n" \
+              f" ├─ New: {self.added_makefiles}, Missing: {self.missing_makefiles}\n" \
+              f" └─ Fail->Success: {self.fail_to_success}, Success->Fail: {self.success_to_fail}"
+        return msg
 
 
 def main():
@@ -326,7 +403,7 @@ def main():
             libraries = set(f.read().split())
 
     try:
-        iterator = iter_repos(db, args.repo_list_file)
+        iterator = iter_repos(db, args.repo_list_file, args.max_repos if args.max_repos != -1 else None)
         pipeline_fn = functools.partial(
             clone_and_compile,
             clone_folder=args.clone_folder, binary_folder=args.binary_folder, archive_folder=args.archive_folder,
@@ -335,10 +412,9 @@ def main():
             force_reclone=args.force_reclone, force_recompile=args.force_recompile,
             docker_batch_compile=args.docker_batch_compile,
             max_archive_size=args.max_archive_size, compression_type=args.compression_type,
-            record_libraries=(args.record_libraries is not None))
+            record_libraries=(args.record_libraries is not None), record_metainfo=args.record_metainfo)
         repo_count = 0
-        cloned_count = 0
-        # for result in map(pipeline_fn, iterator):
+        meta_info = MetaInfo()
         for result in pool.imap_unordered(pipeline_fn, iterator):
             repo_count += 1
             if repo_count % 100 == 0:
@@ -346,29 +422,36 @@ def main():
             if result is None:
                 continue
             result: PipelineResult
-            repo_owner, repo_name = result.repo_owner, result.repo_name
-            if result.clone_success or result.makefiles:
-                cloned_count += 1
-            if result.clone_success is not None:
-                repo_size = result.repo_size or -1  # a value of zero is probably also wrong
-                db.add_repo(repo_owner, repo_name, result.clone_success, repo_size=repo_size)
-                ghcc.log(f"Added {repo_owner}/{repo_name} to DB")
-            if result.makefiles is not None:
-                db.update_makefile(repo_owner, repo_name, result.makefiles, ignore_length_mismatch=args.force_recompile)
+            repo_owner, repo_name = result.repo_info.repo_owner, result.repo_info.repo_name
+            if args.write_db:
+                if result.clone_success is not None or result.repo_info.db_result is None:
+                    # There's probably an inconsistency somewhere if we didn't clone while `db_result` is None.
+                    # To prevent more errors, just add it to the DB.
+                    repo_size = result.repo_size or -1  # a value of zero is probably also wrong
+                    db.add_repo(repo_owner, repo_name, result.clone_success or True, repo_size=repo_size)
+                    ghcc.log(f"Added {repo_owner}/{repo_name} to DB")
+                if result.makefiles is not None:
+                    db.update_makefile(repo_owner, repo_name, result.makefiles,
+                                       ignore_length_mismatch=args.force_recompile)
             if result.libraries is not None:
                 libraries.update(result.libraries)
                 if repo_count % 10 == 0:  # flush every 10 repos
                     with open(args.record_libraries, "w") as f:
                         f.write("\n".join(libraries))
-            if cloned_count >= args.max_repos > 0:
-                break
+
+            if args.record_metainfo:
+                meta_info.add_repo(result)
+                if repo_count % 100 == 0:
+                    ghcc.log(repr(meta_info), force_console=True)
 
     except KeyboardInterrupt:
         print("Press Ctrl-C again to force terminate...")
     except BlockingIOError as e:
-        print(e)
+        print(traceback.format_exc())
         pool.close()
         pool.terminate()
+    except Exception as e:
+        print(traceback.format_exc())
     finally:
         ghcc.log("Gracefully shutting down...", "warning", force_console=True)
         if args.record_libraries is not None:
